refactor: move MetadataApiService modules to an api folder
